Restructure Repo
Added all notebooks to a seperaate folder named "notebooks"
in order to cleanup the repo.

Renamed notebooks in the order of execution i.e. "1." means that
the notebook was created and executed first. "1.1" means that
the notebooks was part of the greater "1." part and followed right
after.

Notebooks needed to be edited in order to be able to import libraries
in the parent directory, which will be done in the next commit.

Finally, added a RemovePunctationDecorator